+++ R/get_yhat.R
@@ -7,7 +7,7 @@
 #' \code{get.yhat()} is designed to always return a simple numeric vector of model predictions.
 #' The design of \code{get.yhat()} is strongly influenced by \code{DALEX::yhat()}.
 #'
-#' @param X.model a fitted model object.
+#' @param object a fitted model object.
 #' @param newdata a data.frame or matrix.
 #' @param target an integer or character vector specifying the target levels for the prediction, used for the models that returns a matrix or data.frame of class probabilities. Default is \code{-1}, representing the probability of not being the base level.
 #' @param ... optional parameters that are passed to the prediction method for the model.
@@ -20,15 +20,15 @@
 #' \code{get.yhat()} returns a numeric vector of model predictions for the \code{newdata}.
 #' @export get.yhat
 #'
-get.yhat <- function(X.model, newdata, ...)
+get.yhat <- function(object, newdata, ...)
 UseMethod("get.yhat")
 
 
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.default <- function(X.model, newdata, target = -1L, ...) {
-  yhat <- stats::predict(X.model, newdata, ...)
+get.yhat.default <- function(object, newdata, target = -1L, ...) {
+  yhat <- stats::predict(object, newdata, ...)
   if ((is.matrix(yhat) || is.data.frame(yhat)) && ncol(yhat) > 1L)
     yhat <- rowSums(yhat[, target, drop = FALSE])
   as.numeric(yhat)
@@ -38,8 +38,8 @@ get.yhat.default <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.mid <- function(X.model, newdata, ...) {
-  yhat <- stats::predict(object = X.model, newdata = newdata,
+get.yhat.mid <- function(object, newdata, ...) {
+  yhat <- stats::predict(object = object, newdata = newdata,
                          type = "response", na.action = "na.pass", ...)
   as.numeric(yhat)
 }
@@ -48,8 +48,8 @@ get.yhat.mid <- function(X.model, newdata, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.lm <- function(X.model, newdata, ...) {
-  yhat <- stats::predict(object = X.model, newdata = newdata,
+get.yhat.lm <- function(object, newdata, ...) {
+  yhat <- stats::predict(object = object, newdata = newdata,
                          type = "response", na.action = "na.pass", ...)
   as.numeric(yhat)
 }
@@ -58,8 +58,8 @@ get.yhat.lm <- function(X.model, newdata, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.glm <- function(X.model, newdata, ...) {
-  yhat <- stats::predict(object = X.model, newdata = newdata,
+get.yhat.glm <- function(object, newdata, ...) {
+  yhat <- stats::predict(object = object, newdata = newdata,
                          type = "response", na.action = "na.pass", ...)
   as.numeric(yhat)
 }
@@ -68,15 +68,15 @@ get.yhat.glm <- function(X.model, newdata, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.rpart <- function(X.model, newdata, target = -1L, ...) {
-  nclass <- length(attr(X.model, "ylevels"))
+get.yhat.rpart <- function(object, newdata, target = -1L, ...) {
+  nclass <- length(attr(object, "ylevels"))
   if (nclass > 0L) {
-    yhat <- stats::predict(object = X.model, newdata = newdata,
+    yhat <- stats::predict(object = object, newdata = newdata,
                            type = "prob", na.action = "na.pass", ...)
     if (is.matrix(yhat) && ncol(yhat) > 1L)
       yhat <- rowSums(yhat[, target, drop = FALSE])
   } else {
-    yhat <- stats::predict(object = X.model, newdata = newdata,
+    yhat <- stats::predict(object = object, newdata = newdata,
                            type = "vector", na.action = "na.pass", ...)
   }
   as.numeric(yhat)
@@ -86,11 +86,11 @@ get.yhat.rpart <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.randomForest <- function(X.model, newdata, target = -1L, ...) {
-  if (X.model$type == "regression") {
-    yhat <- stats::predict(object = X.model, newdata = newdata, ...)
+get.yhat.randomForest <- function(object, newdata, target = -1L, ...) {
+  if (object$type == "regression") {
+    yhat <- stats::predict(object = object, newdata = newdata, ...)
   } else {
-    yhat <- stats::predict(object = X.model, newdata = newdata,
+    yhat <- stats::predict(object = object, newdata = newdata,
                            type = "prob", ...)
     if (is.matrix(yhat) && ncol(yhat) > 1L)
       yhat <- rowSums(yhat[, target, drop = FALSE])
@@ -102,22 +102,16 @@ get.yhat.randomForest <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.ranger <- function(X.model, newdata, target = -1L, ...) {
-  if (na.exists <- anyNA(newdata)) {
-    newdata <- stats::na.exclude(newdata)
-    naa <- stats::na.action(newdata)
-  }
-  if (X.model$treetype == "Regression") {
-    yhat <- stats::predict(object = X.model, data = newdata, ...)$predictions
-  } else if (X.model$treetype == "Probability estimation") {
-    yhat <- stats::predict(object = X.model, data = newdata, ...)$predictions
+get.yhat.ranger <- function(object, newdata, target = -1L, ...) {
+  if (object$treetype == "Regression") {
+    yhat <- stats::predict(object = object, data = newdata, ...)$predictions
+  } else if (object$treetype == "Probability estimation") {
+    yhat <- stats::predict(object = object, data = newdata, ...)$predictions
     if (is.matrix(yhat) && ncol(yhat) > 1L)
       yhat <- rowSums(yhat[, target, drop = FALSE])
   } else {
     stop("'treetype' must be 'Regression' or 'Probability estimation'")
   }
-  if (na.exists)
-    yhat <- stats::napredict(naa, yhat)
   as.numeric(yhat)
 }
 
@@ -125,17 +119,17 @@ get.yhat.ranger <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.svm <- function(X.model, newdata, target = -1L, ...) {
-  if (X.model$type == 0) {
+get.yhat.svm <- function(object, newdata, target = -1L, ...) {
+  if (object$type == 0) {
     yhat <- stats::predict(
-      object = X.model, newdata = newdata, probability = TRUE,
+      object = object, newdata = newdata, probability = TRUE,
       na.action = stats::na.exclude, ...
     )
     yhat <- attr(yhat, "probabilities")
     if (is.matrix(yhat) && ncol(yhat) > 1L)
       yhat <- rowSums(yhat[, target, drop = FALSE])
   } else {
-    yhat <- stats::predict(object = X.model, newdata = newdata,
+    yhat <- stats::predict(object = object, newdata = newdata,
                            na.action = stats::na.exclude, ...)
   }
   as.numeric(yhat)
@@ -145,12 +139,16 @@ get.yhat.svm <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.ksvm <- function(X.model, newdata, target = -1L, ...) {
+get.yhat.ksvm <- function(object, newdata, target = -1L, ...) {
   dots <- list(...)
-  args <- list(object = X.model, newdata = newdata)
-  if (any(X.model@type == c("eps-svr", "eps-bsvr", "nu-svr"))) {
+  if (na.exists <- anyNA(newdata)) {
+    newdata <- stats::na.exclude(newdata)
+    naa <- stats::na.action(newdata)
+  }
+  args <- list(object = object, newdata = newdata)
+  if (any(object@type == c("eps-svr", "eps-bsvr", "nu-svr"))) {
     args$type <- "response"
-  } else if (any(X.model@type == c("C-svc", "nu-svc", "C-bsvm", "spoc-svc"))) {
+  } else if (any(object@type == c("C-svc", "nu-svc", "C-bsvm", "spoc-svc"))) {
     args$type <- "probabilities"
   }
   for (item in names(dots)) {
@@ -161,6 +159,8 @@ get.yhat.ksvm <- function(X.model, newdata, target = -1L, ...) {
   yhat <- do.call("predict", args = args)
   if (is.matrix(yhat) && ncol(yhat) > 1L)
     yhat <- rowSums(yhat[, target, drop = FALSE])
+  if (na.exists)
+    yhat <- stats::napredict(naa, yhat)
   as.numeric(yhat)
 }
 
@@ -168,16 +168,16 @@ get.yhat.ksvm <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.AccurateGLM <- function(X.model, newdata, ...) {
+get.yhat.AccurateGLM <- function(object, newdata, ...) {
   dots <- list(...)
   if (na.exists <- anyNA(newdata)) {
     newdata <- stats::na.exclude(newdata)
     naa <- stats::na.action(newdata)
   }
-  predvars <- sapply(X.model@vars_info, function(x) x$name)
+  predvars <- sapply(object@vars_info, function(x) x$name)
   if (length(setdiff(colnames(newdata), predvars)) > 0L)
     newdata <- newdata[, predvars, drop = FALSE]
-  args <- list(object = X.model, newx = newdata, type = "response")
+  args <- list(object = object, newx = newdata, type = "response")
   for (item in names(dots)) {
     if (item %in% c("object", "newx", "type"))
       next
@@ -195,7 +195,7 @@ get.yhat.AccurateGLM <- function(X.model, newdata, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.glmnet <- function(X.model, newdata, ...) {
+get.yhat.glmnet <- function(object, newdata, ...) {
   dots <- list(...)
   if (na.exists <- anyNA(newdata)) {
     newdata <- stats::na.exclude(newdata)
@@ -203,7 +203,7 @@ get.yhat.glmnet <- function(X.model, newdata, ...) {
   }
   if (!is.matrix(newdata))
     newdata <- as.matrix(newdata)
-  args <- list(object = X.model, newx = newdata, type = "response")
+  args <- list(object = object, newx = newdata, type = "response")
   for (item in names(dots)) {
     if (item %in% c("object", "newx", "type"))
       next
@@ -221,16 +221,16 @@ get.yhat.glmnet <- function(X.model, newdata, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.model_fit <- function(X.model, newdata, target = -1L, ...) {
-  if (X.model$spec$mode == "regression") {
-    yhat <- stats::predict(object = X.model, new_data = newdata,
+get.yhat.model_fit <- function(object, newdata, target = -1L, ...) {
+  if (object$spec$mode == "regression") {
+    yhat <- stats::predict(object = object, new_data = newdata,
                            type = "numeric")
     yhat <- yhat$.pred
-  } else if (X.model$spec$mode == "classification") {
-    yhat <- stats::predict(object = X.model, new_data = newdata, type = "prob")
+  } else if (object$spec$mode == "classification") {
+    yhat <- stats::predict(object = object, new_data = newdata, type = "prob")
     yhat <- as.matrix(yhat)
     if (is.matrix(yhat) && ncol(yhat) > 1L) {
-      colnames(yhat) <- X.model$lvl
+      colnames(yhat) <- object$lvl
       yhat <- rowSums(yhat[, target, drop = FALSE])
     }
   }
@@ -241,16 +241,16 @@ get.yhat.model_fit <- function(X.model, newdata, target = -1L, ...) {
 #' @rdname get.yhat
 #' @exportS3Method midr::get.yhat
 #'
-get.yhat.rpf <- function(X.model, newdata, target = -1L, ...) {
-  if (X.model$mode == "regression") {
-    yhat <- stats::predict(object = X.model, new_data = newdata,
+get.yhat.rpf <- function(object, newdata, target = -1L, ...) {
+  if (object$mode == "regression") {
+    yhat <- stats::predict(object = object, new_data = newdata,
                            type = "numeric")
     yhat <- yhat$.pred
-  } else if (X.model$mode == "classification") {
-    yhat <- stats::predict(object = X.model, new_data = newdata, type = "prob")
+  } else if (object$mode == "classification") {
+    yhat <- stats::predict(object = object, new_data = newdata, type = "prob")
     yhat <- as.matrix(yhat)
     if (is.matrix(yhat) && ncol(yhat) > 1L) {
-      colnames(yhat) <- X.model$lvl
+      colnames(yhat) <- object$lvl
       yhat <- rowSums(yhat[, target, drop = FALSE])
     }
   }