+++ tests/testthat/test-interpret.R
@@ -30,7 +30,7 @@ test_that("interpret() returns weighted least-norm solutions", {
 test_that("interpret() returns valid 'na.action'", {
   x <- c(1, 2, 3, 4, NA)
   y <- c(1, 2, NA, 4, 5)
-  X <- data.frame(x = x, z = c(2, 2, 2, NA, 2))
+  X <- data.frame(x = x, y = c(2, 2, 2, NA, 2))
   # test 1 : formula, vector x
   mid <- interpret(formula = y ~ x)
   expect_equal(mid$na.action, c(3, 5), ignore_attr = TRUE)
@@ -75,4 +75,27 @@ test_that("interpret() returns valid result with weighted data", {
   mid2 <- interpret(x = 1:3, y = (1:3) * 2, weights = c(2, 2, 1))
   expect_equal(mid1$main.effects, mid2$main.effects)
   expect_equal(mid1$intercept, mid2$intercept)
+})
+
+test_that("weights works", {
+  # test 1
+  x1 <- c(1, 1, 1, 1, 2, 2, 2)
+  x2 <- c(1, 1, 1, 2, 1, 2, 2)
+  x3 <- c(1, 1, 2, 1, 2, 1, 2)
+  X1 <- data.frame(cbind(x1, x2, x3))
+  x1 <- c(1, 1, 1, 2, 2, 2)
+  x2 <- c(1, 1, 2, 1, 2, 2)
+  x3 <- c(1, 2, 1, 2, 1, 2)
+  weights <- c(2, 1, 1, 1, 1, 1)
+  X2 <- data.frame(cbind(x1, x2, x3))
+  X3 <- weighted(X2, weights)
+  r1 <- interpret(x1 * x2 * x3 ~ (x1 + x2 + x3), X1)$ratio
+  r2 <- interpret(x1 * x2 * x3 ~ (x1 + x2 + x3), X2)$ratio
+  r3 <- interpret(x1 * x2 * x3 ~ (x1 + x2 + x3), X2, weights = weights)$ratio
+  r4 <- interpret(x1 * x2 * x3 ~ (x1 + x2 + x3), X3)$ratio
+  r5 <- interpret(x1 * x2 * x3 ~ (x1 + x2 + x3), X3, weights = rep(1, 6))$ratio
+  expect_equal(r1, r3)
+  expect_equal(r1, r4)
+  expect_equal(r2, r5)
+  expect_false(r1 == r2)
 })